🐛 fix: build error 해결을 위한 폴더구조 수정
