+++ src/components/main/styled.tsx
